feat(clap): autogen bash completions
