+++ Cargo.toml
@@ -10,6 +10,7 @@ rust-version = "1.65.0"
 readme = "README.md"
 license = "GPL-3.0-only"
 version = "0.1.1"
+build = "build.rs"
 
 [dependencies]
 clap = { version = "4.4.6", features = ["cargo"] }
@@ -22,3 +23,7 @@ rand = "0.8.5"
 [dev-dependencies]
 assert_cmd = "2.0.12"
 tempfile = "3.8.0"
+
+[build-dependencies]
+clap = { version = "4.4.6", features = ["cargo"] }
+clap_complete = "4"
