+++ src/cli.rs
@@ -1,0 +1,33 @@
+use clap::{Command, arg, command, crate_authors, Arg};
+
+pub fn build_cli() -> Command {
+    command!()
+        .author(crate_authors!("\n"))
+        .arg(
+            Arg::new("all")
+                .short('a')
+                .long("all")
+                .help("Shows all fortunes, including unkind."),
+        )
+        .arg(
+            Arg::new("unkind")
+                .short('o')
+                .short('u')
+                .long("unkind")
+                .help("Shows only unkind fortunes."),
+        )
+        .arg(
+            Arg::new("find")
+                .short('m')
+                .long("find")
+                .value_name("pattern")
+                .help("Finds fortunes matching regex query."),
+        )
+        .arg(
+            Arg::new("length")
+                .short('n')
+                .long("length")
+                .help("Finds a fortune that is shorter than provided number."),
+        )
+        .arg(arg!(-s --short ... "Shows a short aporism."))
+}
