+++ src/main.rs
@@ -2,40 +2,11 @@ use clap::{Command, arg, command, crate_authors, Arg};
 use std::env;
 use std::io;
 
-pub fn build_cli() -> Command {
-    command!()
-        .author(crate_authors!("\n"))
-        .arg(
-            Arg::new("all")
-                .short('a')
-                .long("all")
-                .help("Shows all fortunes, including unkind."),
-        )
-        .arg(
-            Arg::new("unkind")
-                .short('o')
-                .short('u')
-                .long("unkind")
-                .help("Shows only unkind fortunes."),
-        )
-        .arg(
-            Arg::new("find")
-                .short('m')
-                .long("find")
-                .value_name("pattern")
-                .help("Finds fortunes matching regex query."),
-        )
-        .arg(
-            Arg::new("length")
-                .short('n')
-                .long("length")
-                .help("Finds a fortune that is shorter than provided number."),
-        )
-        .arg(arg!(-s --short ... "Shows a short aporism."))
-}
+mod cli;
 
 fn main() -> io::Result<()> {
-    let matches = build_cli().get_matches();
+    let matches = cli::build_cli().get_matches();
+
     if let Some(pattern) = matches.get_one::<String>("find") {
         fortune::search_fortunes(pattern);
     } else if let Some(short) = matches.get_one::<u8>("short") {
